+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - **BREAKING** The method `Youthweb\UrlLinker\UrlLinker::getAllowFtpAddresses()` was removed
 - **BREAKING** The method `Youthweb\UrlLinker\UrlLinker::setAllowFtpAddresses()` was removed
+- **BREAKING** The method `Youthweb\UrlLinker\UrlLinker::getAllowUpperCaseUrlSchemes()` was removed
+- **BREAKING** The method `Youthweb\UrlLinker\UrlLinker::setAllowUpperCaseUrlSchemes()` was removed
 
 ## [1.5.0](https://github.com/youthweb/urllinker/compare/1.4.0...1.5.0) - 2022-12-07